Passing SliderCursors-interface to sliders
Passing an interface to sliders instead of having them
get_art_container().

Not entirely happy how this turned out. It is too complex now,
especially using a "Getter" std::function.

At least it is a step on the way to breaking the foul get_art_containers
hold on the application.