Import * from ifaint into faint
The ifaint (faint builtins) module name is mostly an accident of
implementation, all faint-code should be in the faint-package.

I achieved this (for now?) by importing all of ifaint into faint
by from ifaint import * in the faint __init__.py.

I also renamed the extension (ext/) to ifaint, so that it can be imported
in the same way from py/ and coexist with the faint package (.py files).

Also: I catch ImportError from the import for now, to allow the old
tests in py/test to run, since some of them do not run through faint, nor
use the extension module so they lack ifaint.