Avoid assertion on Enter-key for selected-color-ctrl.
Faint failed an assertion in selected-color-ctrl.cpp for these steps:
1. Start Faint
2. Click a color in the selected color control (not the palette!)
3. Close the dialog with enter
4. Hit enter again

This asserted because it passes some unexpected event id to
HandleMenuChoice, which calls MenuTargetColor while m_menuEventColor
is initially which::HIT_NEITHER, which is not allowed.

I reorganized the logic to only initialize the color for known events,
this adds some duplication but makes the conditions self-contained,
reduces nesting and, especially, avoids calling a function that
crashes.

I'm not sure why I get a menu event in this situation though.